blogPostOne
24JUN2024
i got easyOS to run from a USB. it's a hassle to set up, the session information is never actually saved to the drive, but it actually feels nice to use.
i'd say easyOS is on par with linuxMint tbh. it certainly feels <i>different</i> but not any better or worse. it's just the setup process and inability to save.
the reason im using a USB btw is bc i don't just have an empty hardDrive laying around. i could just wipe one of the drives though - i don't have anything important.</p>